requirements.txt
# Core data processing
polars>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# API & networking
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Statistical models
statsmodels>=0.14.0

# Visualization (for notebooks)
matplotlib>=3.8.0
seaborn>=0.13.0

# Bayesian inference (Phase 2)
# pymc>=5.10.0
# numpyro>=0.14.0
# arviz>=0.17.0

# Testing
pytest>=7.4.0
